requirements.txt
python-dotenv==1.0.1
sounddevice==0.5.1
soundfile==0.13.0
